Move Header component's location
